Reduce code duplication
Move generic code to common.go to reduce code duplication.
